SNES: read the X / Y / L / R buttons from scripts
The input opcodes (IF_INPUT, AWAIT_INPUT, SET_INPUT_SCRIPT,
REMOVE_INPUT_SCRIPT) carried a 1-byte button mask, full with the 8 Game Boy
buttons. On the SNES target they now carry a 2-byte little-endian mask - the
extra byte holds X/Y/L/R (KEY_BITS bits 8..11).

Compiler:
- helpers.js KEY_BITS gains x/y/l/r (0x100..0x800)
- targets/{gb,snes}.js: new `inputMaskBytes` (1 / 2), guarded by targets.test
- scriptBuilder.js: new inputMask() helper reads it; the 4 input methods use it
  instead of pushing inputDec() directly. GB path emits the exact same byte.

SNES engine (appData/src/snes/):
- script_cmds.c: those 4 opcodes' args_len +1 (the only divergence from the GB
  table; snesScriptCmds.test allows exactly this). Handlers read
  args[0] | args[1]<<8; jump/index args shift by one.
- scene.c: SceneGbInputBits() -> u16, packs all 12 buttons; input_script_ptrs[]
  -> NUM_INPUT_SCRIPTS (12); loop bounds follow.
- script_runner.{c,h}: await_input -> u16.

Editor: InputPicker.js is Redux-connected and shows a third X/Y/L/R row when
settings.target === "snes". Web player already wired X/Y/L/R (keys u/i/o/p).

The Game Boy engine, its byte-exact event tests and GB ROM output are
untouched. Verified in Mesen: a scene registering SET_INPUT_SCRIPT("x") /
SET_INPUT_SCRIPT("r") had both sub-scripts fire on the real presses (marker
vars read back 42 / 43 from WRAM).

+++ appData/src/snes/src/scene.h
@@ -69,13 +69,14 @@ s16 SceneActorTileY(u8 i);
 
 /* M7-cont.: SET_INPUT_SCRIPT / SET_TIMER_SCRIPT execution. `joy`/`prev_joy`
  * are raw PVSnesLib pad bits; the compiler emits masks in the GB engine's
- * compact 8-bit layout (KEY_BITS in src/lib/compiler/helpers.js), so
- * SceneGbInputBits() re-packs the SNES bits into that layout wherever a
- * compiler-emitted mask needs comparing against - also used by IF_INPUT. */
-u8 SceneGbInputBits(u16 j);
+ * compact button layout (KEY_BITS in src/lib/compiler/helpers.js) - 8 bits on
+ * GB, plus X/Y/L/R in bits 8..11 on SNES - so SceneGbInputBits() re-packs the
+ * SNES bits into that layout wherever a compiler-emitted mask needs comparing
+ * against - also used by IF_INPUT / AWAIT_INPUT. */
+u16 SceneGbInputBits(u16 j);
 void SceneScheduledScriptsInit(void); /* boot-only: zero input_script_ptrs[] */
-void SceneSetInputScript(u8 mask, BANK_PTR target);
-void SceneRemoveInputScript(u8 mask);
+void SceneSetInputScript(u16 mask, BANK_PTR target);
+void SceneRemoveInputScript(u16 mask);
 void SceneSetTimerScript(u8 duration, BANK_PTR target);
 void SceneTimerRestart(void);
 void SceneTimerDisable(void);
